Prevent Auth Provider side-nav navigation if it's a no-op
- Addresses #2321
- When there's no auth provider clicking on the Auth Provider side nav menu item takes the user to the list of possible auth providers
- When an auth provider is enabled clicking on the same button redirects the user to the detail of that auth provider
- When already on that detail page clicking again on the auth provider item resulted in the nuxt page loading component to be shown and never be removed
- Now we prevent navigation away from this page if we're going back to the same place
- Note - I've tried a number of different ways of fixing this in the root auth/config page without joy (using a page guard directly within `middleware` and also more traditional guards with beforeResolve and beforeRouteEnter). Although more code this does work and visually clicking the menu item looks the same as on any other page.
