+++ pages/c/_cluster/auth/config/_id.vue
@@ -1,6 +1,7 @@
 <script>
 import ResourceDetail from '@/components/ResourceDetail';
 import { MANAGEMENT } from '@/config/types';
+import { authProviders } from '@/pages/c/_cluster/auth/config/auth-config-utils';
 
 export default {
   name:        'AuthConfigDetail',
@@ -8,6 +9,19 @@ export default {
   computed:   {
     AUTH_CONFIG() {
       return MANAGEMENT.AUTH_CONFIG;
+    }
+  },
+  async beforeRouteLeave(to, from, next) {
+    // The root auth/config page redirects to this page in some cases. Prevent that navigation if we're just going to end up here again
+    const sameRoute = from.name === 'c-cluster-auth-config-id' && to.name === 'c-cluster-auth-config';
+
+    if (sameRoute) {
+      // Ensure we re-evaluate the redirect in case this auth provider has been disabled
+      const authProvs = await authProviders(this.$store);
+
+      next(!authProvs.enabledLocation);
+    } else {
+      next();
     }
   },
 };
